refactor: relocate KoColor domain models to core ritual module
This commit updates the import paths for several domain models in the KoColor application, moving them from the local model package to the centralized `com.zoewave.probase.core.model.ritual` package. This aligns the home feature with the project's shared architectural models for rituals, suggestions, and fashion profiles.

- **`:applications:kocolor:mobile:features:home`**:
    - **`HomeViewModel.kt`**: Updated `HomeState` and the ViewModel to use core ritual models including `BeautyRoutine`, `ClothingItem`, `CosmeticItem`, `FashionProfile`, `RoutineTime`, and `SavedAnalysis`.
    - **`CollectionDetailScreen.kt`**: Refactored imports for `FashionAdvice`, `MakeupSuggestion`, `OutfitSuggestion`, `SavedAnalysis`, `SeasonalType`, and `Undertone`.
    - **`HomeScreen.kt` & `CollectionHubScreen.kt`**: Updated references to `FashionProfile` and `SavedAnalysis` to use the centralized core package.
    - **`RoutineSummaryCard.kt`**: Migrated the `BeautyRoutine` reference to the core ritual module.
